requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.5
python-decouple==3.8
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0